Update and rename ose_075.py to ose075.py
